move extract logic to ChapterOutput
